==> 2.CleaningData_2016.Rmd <==
---
title: "Cleaning data 2016"
author: "Camila Calderon"
date: "2024-07-24"
output: html_document
---

### Load libraries

```{r setup, warning=FALSE, message=FALSE}
pacman::p_load(move,tidyverse, move, lubridate, amt, ggplot2)
```

### Download data from movebank from Bocas 2016 

```{r}
me <- movebankLogin(username="myaccount", password="mypassword")#use own credentials
studyId <- getMovebankID("Foraging movements of greater spear-nosed bats, Bocas del Toro",me)
batsmove2016<- getMovebankData(study=studyId, login=me)

# exclude bat with coordinates far away
batsmove2016 <- batsmove2016[batsmove2016$tag_id!="146399774",]

# convert to data frame
batsdf2016 <- as.data.frame(batsmove2016)
unique(batsdf2016$tag_local_identifier)
```

### Remove individuals with incomplete data

```{r, results=FALSE}
# look at the dataframe
head(batsdf2016)

# checking for individuals with complete data
ggplot(aes(x=location_long, y=location_lat, color=tag_local_identifier), data=batsdf2016[batsdf2016$tag_local_identifier=="74D8C25",])+geom_point() #leave

ggplot(aes(x=location_long, y=location_lat, color=tag_local_identifier), data=batsdf2016[batsdf2016$tag_local_identifier=="74D972D",])+geom_point() #remove

ggplot(aes(x=location_long, y=location_lat, color=tag_local_identifier), data=batsdf2016[batsdf2016$tag_local_identifier=="74DAF9C",])+geom_point()#leave

ggplot(aes(x=location_long, y=location_lat, color=tag_local_identifier), data=batsdf2016[batsdf2016$tag_local_identifier=="74F7D4C",])+geom_point()#remove

# add date
batsdf2016$date <- date(batsdf2016$timestamps)

# remove individuals with incomplete data for 2016
batsdf2016<- batsdf2016[!(batsdf2016$tag_local_identifier=="74F7D4C"  |batsdf2016$tag_local_identifier== "74D972D"),] 

# individual "74F7D4C" is foraging in isla colon and I am not sure if the track is complete. Thus, I've removed it.
```

### Downsample data to every 2 minutes

```{r, results=FALSE}
# order by ID and timestamps
batsdf2016 <- batsdf2016[order(batsdf2016$tag_local_identifier, batsdf2016$timestamp),]

# add ID column
batsdf2016$ID<- paste(batsdf2016$tag_local_identifier, date(batsdf2016$timestamp), sep="_")

# split by ID
bats_2016_split <- split(batsdf2016, batsdf2016$ID)

# convert to track object before downsampling
resample <-  lapply(bats_2016_split, function(x){
  mk_track(tbl=x, all_cols=T,
           .x=location_long, .y=location_lat, crs = 32617,
           .t=timestamp, order_by_ts = T, check_duplicates = T)
})

# downsample every two minutes
resample_2 <- lapply(resample, function(x){
  track_resample(x, rate = minutes(2), tolerance = seconds(1), start = 1)
})

# check if resampling went well
lapply(resample_2, function(x){
  ggplot(aes(x=x_, y=y_), data=x) + geom_point()+
    ggtitle(x$ID)
})

# remove some bad data
resample_2 <- resample_2[names(resample_2) %in% c("2016030703_2016-03-10","74DA035_2016-03-04", "74DC454_2016-03-04", "74DDA80_2016-03-10", "74DDFB1_2016-03-04", "74F9F83_2016-03-06", "74FE24E_2016-03-06") ==FALSE]

# convert to data frame
batsdf2016 <- do.call(rbind.data.frame,resample_2)
batsdf2016 <- as.data.frame(batsdf2016)
names(batsdf2016)[1] <- "location_long"
names(batsdf2016)[2] <- "location_lat"
names(batsdf2016)[3] <- "timestamp"

# convert again to move object to calculate movement variables
phas2016move <-  move(x=batsdf2016$location_long, y=batsdf2016$location_lat, 
     time=as.POSIXct(batsdf2016$timestamp, format="%Y-%m-%d %H:%M:%S", tz="UTC"), 
     data=batsdf2016, proj=CRS("+proj=longlat +ellps=WGS84"), 
     animal=batsdf2016$tag_local_identifier, sensor="GPS")

# add movement parameters and batday
moveList <- lapply(split(phas2016move), function(myInd){
  datechange <- c(0, abs(diff(as.numeric(as.factor(date(myInd@timestamps-(12*60*60)))))))
  myInd$BatDay <- cumsum(datechange)+1
  myInd$tlag <- c(NA, move::timeLag(myInd, units="secs"))
  myInd$step <- c(NA, move::distance(myInd))
  myInd$tag_ground_speed <- myInd$ground_speed
  myInd$ground_speed <- c(NA, move::speed(myInd))
  myInd$angle <- c(NA, angle(myInd))
  return(myInd)
})

phas <- moveStack(moveList, forceTz="UTC")

phas <- spTransform(phas, CRS("+proj=utm +zone=17 +datum=WGS84"))
crds <- as.data.frame(phas@coords)
phas$x <- crds$coords.x1 #27
phas$y <- crds$coords.x2 #28

# convert move object to dataframe
phasprocessed <- as.data.frame(phas)

# add id_batday column 
phasprocessed$ID_batday <- paste(phasprocessed$tag_local_identifier, phasprocessed$BatDay, sep="_")

# filter only females
bats_2016 <- phasprocessed %>%
  filter(sex=="f")

# save final data frame
save(bats_2016, file="~/ownCloud/PhDLife/P.hastatus/Thesis/Paper2/PlosOne/data/bats2016_resample_clean.RData")
```
