==> 9.Activity_budgets.Rmd <==
---
title: "Activity budgets"
author: "Camila Calderon"
date: "2024-07-30"
output: html_document
---

### Load packages

```{r setup}
pacman::p_load(lubridate, dplyr, ggplot2, tidyverse, plyr, Rmisc)
```

### Loading data

```{r}
load(file="~/ownCloud/PhDLife/P.hastatus/Thesis/paper2/analysis/data/females_HMMbehaviors.RData")

# check cave column
unique(females_bocas$cave)
```

### Plot tracks

```{r, results='hide', message=FALSE}

# split dataframe in a list
females_ls <- split(females_bocas, females_bocas$ID)

# plot all females to check which have complete tracks
pdf("~/ownCloud/PhDLife/P.hastatus/Thesis/Paper2/analysis/results/tracks_females.pdf")
lapply(females_ls, function(x){
  ggplot(aes(x=location_long, y=location_lat, col=behav), data=x)+
  geom_point()+
  ggtitle(unique(x$ID))
})
dev.off()

```


### Remove ids

I am removing individuals with incomplete commutes

```{r}
# remove some id-days
females_com <- females_ls[names(females_ls) %in% c("74F8E19_2016-03-07", "2016030705_2016-03-08", "22507B4E_D_2022-02-01", "74DA035_2016-03-04", "74DAF9C_2016-02-29", "74DC454_2016-03-04", "74DDA80_2016-03-10", "74DDFB1_2016-03-04", "74F9F83_2016-03-06", "74FE24E_2016_03_06") == FALSE] 

# convert list to dataframe
females_com <- do.call(rbind.data.frame,females_com)

# for foraging points which have tlag larger than 1800 sec (30 minutes), because the bat was at the cave, and the gps did not record. We will assign a tlag of 190
females_com$tlag[which(females_com$behav=="foraging" & females_com$tlag>1800)] <- 190
females_com <- females_com %>%
  filter(!c(behav=="commuting" & tlag>1800))
```

### Calculate proportion of time spent in each behavior

```{r, message=FALSE}
# summary data frame of proportions
behav.time= aggregate(females_com[c("tlag")], by=females_com[c("BatDay", "year_cave","tag_local_identifier", "behav", "cave")], FUN=sum, na.rm=T)
behav.time$time.min <- behav.time$tlag/60
behav.time$IDday <- paste(behav.time$tag_local_identifier, behav.time$BatDay, sep="_")
behav.time<- behav.time %>%
  group_by(IDday) %>%
  dplyr::mutate(totaltime=sum(time.min))#time.all$timeTrack.min[match(time.behav.group$IDday, time.all$IDday)]
behav.time$prop.time <- round(behav.time$time.min/behav.time$totaltime,2)

# time spent on each behavior by cave/year
sumtimeBycave<- behav.time %>%  
  group_by(behav,year_cave)%>%# to make summary table by behaviors
  dplyr::summarise(mean.time=mean(time.min), sd.time=sd(time.min))

sumpropBycave<- behav.time %>% 
  group_by(behav, cave)%>%
  dplyr::summarise(mean.prop=mean(prop.time), sd.prop=sd(prop.time))

# number individuals to the summary table
x <- as.data.frame(table(behav.time$year_cave,behav.time$tag_local_identifier))
z <- count(x$Var1[which(x$Freq>0)])

```

###  Plot proportion of time spent on each behavior by cave/year

```{r, echo=FALSE, message=FALSE, warning=FALSE}
# levels of the cave groups
level_order <- c("2016_lagruta", "2022_lagruta_Feb") 

pos <- position_dodge(0.9)

# plot Fig S2
ggplot(aes(x=factor(year_cave, level=level_order), y=prop.time, fill=behav), data=behav.time)+
  #geom_violin(position = pos) +
  geom_boxplot(width = .2, 
               fatten = NULL,
               #alpha = 0.75,
               position = pos)+
  stat_summary(fun = "mean",  geom = "point", position = pos) +
  stat_summary(fun.data = "mean_se", 
               geom = "errorbar", 
               width = .1, position = pos)+
  scale_x_discrete(labels=str_wrap(c("2016",  "2022"), width=10))+
  scale_fill_brewer(palette = "Greys")+
  ylab("proportion of time")+
  xlab("years")+
  labs(fill="behaviors")+
  theme_classic()+
  theme(axis.title = element_text(face = 'bold', size=20), 
        axis.title.x = element_text(size = 24), 
        axis.title.y=element_text(size=24), 
        axis.text.y = element_text(size=20), 
        axis.text.x = element_text(size=20),
        legend.title =element_text(size=20),
        legend.text = element_text(size=18)
        )

# save plot
ggsave(file = "~/ownCloud/PhDLife/P.hastatus/Thesis/Paper2/PlosOne/figures/FigS2.pdf", height=10, width=15)
```

### Test differences in behaviors

```{r, message=FALSE, warning=FALSE}
library(lme4)
library(performance)
library(emmeans)
library(car)

# dividing data frame in commuting and foraging
commuting <- behav.time[behav.time$behav=="commuting",]
foraging <- behav.time[behav.time$behav=="foraging",]

#remove "2D507235_G_1"
foraging <- foraging[foraging$IDday=="2D507235_G_1",]

# checking distribution of the data
hist(commuting$prop.time)
qqnorm(commuting$prop.time)
qqline(commuting$prop.time)

# organize data frame by proportion of behaviors
prop.table <- commuting
prop.table$prop.foraging <- foraging$prop.time
  
# testing differences in commuting by group
modcom_year <- glm(prop.time~year_cave, data=commuting, family = binomial)

# summary of the model
summary(modcom_year)
plot(modcom_year)

# plot model of commuting
plot(modcom_year)
emmeans(modcom_year, specs = pairwise ~ year_cave)

# testing differences in foraging by seasons
foraging <- behav.time[behav.time$behav=="foraging",]

# check data distribution
hist(foraging$prop.time)
qqnorm(foraging$prop.time)
qqline(foraging$prop.time)

# testing differences in foraging by cave year
modfor_cave <- glm(prop.time~year_cave, data=foraging, family = binomial)

# summary
summary(modfor_cave)
plot(modfor_cave)
emmeans(modfor_cave, specs = pairwise ~ year_cave)
```

### Calculate timing of departures after sunset

```{r sunset, warning=FALSE, message=FALSE}
library(sp)
# calculate time ranges
timeRanges <- females_com %>% group_by(year_cave, tag_local_identifier,ID) %>%
  dplyr::summarise(minTimeUTC = min(timestamps),
                   maxTimeUTC = max(timestamps),
                   minTimeLocal = min(with_tz(timestamps, tz="America/Panama"), na.rm = TRUE),
                   maxTimeLocal = max(with_tz(timestamps, tz="America/Panama"), na.rm = TRUE),
                   timeTrack.min = round(as.numeric(as.duration(maxTimeLocal-minTimeLocal))/60, 2),
                   nlocs = n(),
                   meanTimeLag.s = mean(tlag, na.rm = TRUE),
                   maxTimeLag.s = max(tlag, na.rm = TRUE))


# get sunrise/sunset times
coords <- cbind(-82.271541, 9.396448)
lagruta <- SpatialPoints(coords, proj4string = CRS("+proj=longlat +datum=WGS84"))
lagrutaUTM <- spTransform(lagruta, CRS("+proj=utm +zone=17 +datum=WGS84"))
library(maptools)
lagrutaLL <- spTransform(lagrutaUTM, CRS("+proj=longlat +datum=WGS84"))
sunset <- sunriset(coordinates(lagrutaLL), timeRanges$minTimeLocal, direction="sunset", POSIXct.out=TRUE)[2]

# difference to sunset
timeRanges$sunset <- sunset
timeRanges$timeSinceSunset <- timeRanges$minTimeLocal - sunset$time
timeRanges$timeSinceSunset[timeRanges$timeSinceSunset<0] <- timeRanges$timeSinceSunset[timeRanges$timeSinceSunset<0]+1440

# total mean
mean(timeRanges$timeSinceSunset) 
sd(timeRanges$timeSinceSunset) 

# summary sunset times by cave year
sum_timeRanges <- timeRanges %>%
  group_by(year_cave)%>%
  dplyr::summarise(mean_time=mean(timeSinceSunset), sd_time=sd(timeSinceSunset), n=n())

# adding number individuals to the summary table
l <- as.data.frame(table(timeRanges$year_cave,timeRanges$tag_local_identifier))
id <- count(x$Var1[which(x$Freq>0)])
sum_timeRanges$n <- z$freq
sum(sum_timeRanges$n)
```

### Plot FigS1 - timing of departures after sunset
```{r, echo=FALSE, warning=FALSE}

ggplot(aes(x=year_cave, y=timeSinceSunset), data=timeRanges) +
geom_boxplot()+
  geom_jitter()+
  stat_summary(fun = mean, color = "red") +
  stat_summary(
    fun.min = function(x) mean(x) - sd(x), 
    fun.max = function(x) mean(x) + sd(x), 
    geom = "errorbar",
    color = "red",
    width = .3
  )+
  scale_x_discrete(labels=c("2016", "2022"))+
  ylab("departure time after sunset (mins)")+
  xlab("years")+
  theme_classic()

ggsave(file="~/ownCloud/PhDLife/P.hastatus/Thesis/Paper2/PlosOne/figures/FigS1.pdf")
```
